% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_series_plot.R
\name{time_series_plot}
\alias{time_series_plot}
\title{Interactive Plotting for One Time Series}
\usage{
time_series_plot(df, date, value)
}
\arguments{
\item{df}{A tibble or data.frame with a time-based column}

\item{date}{A column containing either date or date-time values}

\item{value}{A column containing numeric values}
}
\value{
An interactive plotly plot
}
\description{
Interactive Plotting for One Time Series
}
\examples{
library(ggplot2)
time_series_plot(economics, date, unemploy)
}
